requirements.txt
langchain==0.3.3
langchain-google-genai==2.0.1
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.39.0
faiss-cpu==1.7.4
streamlit-extras==0.3.5